# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 채용공고 Inbox API
  version: 0.3.0
  description: >
    Hand-maintained snapshot of the implemented MVP REST surface.
    Model: recruit-inbox-mvp-design-v1.1.md §7. Frontend types mirror this in
    frontend/lib/types.ts. springdoc auto-generation is deferred until a
    Spring Boot 4 compatible release ships. Production uses the Google/Kakao OAuth2
    session cookie; development may use X-Dev-User-Id. Mutating requests also
    require the X-XSRF-TOKEN header obtained from /api/v1/auth/csrf.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
  - devUser: []
...
paths:
...
  /api/v1/profile-items:
    get:
      tags: [profile]
      summary: List the current user's reusable application information
      parameters:
        - name: category
          in: query
          required: false
          schema: { $ref: "#/components/schemas/ProfileCategory" }
      responses:
        "200":
          description: profile items
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ProfileItemResponse" } }
    post:
      tags: [profile]
      summary: Add a reusable application-information item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileItemCreate" }
      responses:
        "201": { description: created, content: { application/json: { schema: { $ref: "#/components/schemas/ProfileItemResponse" } } } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/profile-items/{id}:
    patch:
      tags: [profile]
      summary: Update an owned profile item
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileItemUpdate" }
      responses:
        "200": { description: updated, content: { application/json: { schema: { $ref: "#/components/schemas/ProfileItemResponse" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [profile]
      summary: Delete an owned profile item
      parameters:
        - { $ref: "#/components/parameters/Id" }
        - { name: If-Match, in: header, required: true, schema: { type: string } }
      responses:
        "204": { description: deleted }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Session cookie established by Google or Kakao OAuth2 login in production.
    devUser:
      type: apiKey
      in: header
      name: X-Dev-User-Id
      description: Development only; the UUID must identify an existing local user.
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  retryable: { type: boolean }
                  requestId: { type: string }
                  fields: { type: object, additionalProperties: { type: string } }
  schemas:
...
    ProfileCategory:
      type: string
      enum: [PERSONAL, MILITARY, EDUCATION, LANGUAGE, CERTIFICATION, CAREER, AWARD, ACTIVITY, SKILL, PROJECT, STORY]
    ProfileItemCreate:
      type: object
      required: [category]
      properties:
        category: { $ref: "#/components/schemas/ProfileCategory" }
        label: { type: string, minLength: 1, maxLength: 200 }
        fields:
          type: object
          description: Category-specific field key/value pairs. Unsupported keys are rejected.
          additionalProperties: { type: string, maxLength: 20000 }
        valueText: { type: string, maxLength: 20000 }
        details: { type: string, maxLength: 100000 }
        startedOn: { type: string, format: date }
        endedOn: { type: string, format: date }
        sensitive: { type: boolean }
    ProfileItemUpdate:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64 }
        category: { $ref: "#/components/schemas/ProfileCategory" }
        label: { type: string, minLength: 1, maxLength: 200 }
        fields:
          type: object
          additionalProperties: { type: string, maxLength: 20000 }
        valueText: { type: string, maxLength: 20000 }
        details: { type: string, maxLength: 100000 }
        startedOn: { type: string, format: date }
        endedOn: { type: string, format: date }
        clearDates: { type: boolean }
        sensitive: { type: boolean }
    ProfileItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        category: { $ref: "#/components/schemas/ProfileCategory" }
        label: { type: string }
        fields:
          type: object
          additionalProperties: { type: string }
        valueText: { type: string }
        details: { type: string }
        startedOn: { type: [string, "null"], format: date }
        endedOn: { type: [string, "null"], format: date }
        sensitive: { type: boolean }
        sortOrder: { type: integer }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...